Update from bencher and merge benchmarks into sync_mpsc.rs
Previous benchmarks in sync_mpsc_recv_many.rs included a
contrived example that is removed.
